Moved wrapped things to 'utils' submodule.
